Track velocity for drag (centroid), scale (span), and rotate (angle).
Summary:
Create new ValueVelocityTracker helper class.
Add velocity getters to gesture recognizers.

Fixes https://github.com/material-motion/family-direct-manipulation-android/issues/15

ValueVelocityTracker was created as a wrapper around VelocityTracker. VelocityTracker reports the velocity of individual pointers, which is not what we need. We need the velocity of the change in centroid, change in span, and change in rotation, which can not be easily derived from the velocities of the individual pointers.

Reviewers: miguelandres, O2 Material Motion, O6 Material Motion Android platform reviewers, featherless

Reviewed By: miguelandres, O2 Material Motion, O6 Material Motion Android platform reviewers, featherless

Subscribers: featherless, miguelandres

Differential Revision: http://codereview.cc/D2007